ThirdModuleOrenburgBackend/serializers.py
from django.contrib.auth.models import AbstractUser, User
from rest_framework import serializers

class RegisterSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'password')